// api/api.ts
export const AuthMethod = {
    Password: 'Password',
    Email: 'Email',
    Mobile: 'Mobile'
} as const;

export type AuthMethod = typeof AuthMethod[keyof typeof AuthMethod];


/**
 * 
 * @export
 * @interface AuthorizeCommand
 */
export interface AuthorizeCommand {
    /**
     * 
     * @type {string}
     * @memberof AuthorizeCommand
     */
    'callbackURL'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AuthorizeCommand
     */
    'clientId'?: string;
}
/**
 * 
 * @export
 * @interface CreateServiceCommand
 */
export interface CreateServiceCommand {
    /**
     * 
     * @type {string}
     * @memberof CreateServiceCommand
     */
    'serviceName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateServiceCommand
     */
    'description'?: string | null;
    /**
     * 
     * @type {ServiceType}
     * @memberof CreateServiceCommand
     */
    'serviceType'?: ServiceType;
    /**
     * 
     * @type {string}
     * @memberof CreateServiceCommand
     */
    'url'?: string | null;
    /**
     * 
     * @type {AuthMethod}
     * @memberof CreateServiceCommand
     */
    'minimumAuthMethod'?: AuthMethod;
    /**
     * 
     * @type {string}
     * @memberof CreateServiceCommand
     */
    'authorizeURL'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateServiceCommand
     */
    'logoutHookURL'?: string | null;
}
// ...
export interface ManagerServiceDto {
    /**
     * 
     * @type {string}
     * @memberof ManagerServiceDto
     */
    'id'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ManagerServiceDto
     */
    'name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ManagerServiceDto
     */
    'description'?: string | null;
    /**
     * 
     * @type {AuthMethod}
     * @memberof ManagerServiceDto
     */
    'authMethod'?: AuthMethod;
    /**
     * 
     * @type {ServiceType}
     * @memberof ManagerServiceDto
     */
    'serviceType'?: ServiceType;
    /**
     * 
     * @type {ServiceState}
     * @memberof ManagerServiceDto
     */
    'serviceState'?: ServiceState;
    /**
     * 
     * @type {string}
     * @memberof ManagerServiceDto
     */
    'organization'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ManagerServiceDto
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof ManagerServiceDto
     */
    'url'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ManagerServiceDto
     */
    'appId'?: string;
    /**
     * 
     * @type {string}
     * @memberof ManagerServiceDto
     */
    'authorizeUrl'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ManagerServiceDto
     */
    'logoutHookUrl'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ManagerServiceDto
     */
    'publicKey'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof ManagerServiceDto
     */
    'feNotificationsEnabled'?: boolean;
}
// ...
export const ServiceState = {
    Registered: 'REGISTERED',
    Pending: 'PENDING'
} as const;

export type ServiceState = typeof ServiceState[keyof typeof ServiceState];


/**
 * 
 * @export
 * @enum {string}
 */

export const ServiceType = {
    Website: 'Website',
    Mobile: 'Mobile',
    Desktop: 'Desktop'
} as const;

export type ServiceType = typeof ServiceType[keyof typeof ServiceType];


/**
 * 
 * @export
 * @interface StatusDto
 */
export interface StatusDto {
    /**
     * 
     * @type {string}
     * @memberof StatusDto
     */
    'message'?: string | null;
}
/**
 * 
 * @export
 * @interface TokenCommand
 */
export interface TokenCommand {
    /**
     * 
     * @type {string}
     * @memberof TokenCommand
     */
    'clientId'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TokenCommand
     */
    'authToken'?: string | null;
}
/**
 * 
 * @export
 * @interface UpdateServiceCommand
 */
export interface UpdateServiceCommand {
    /**
     * 
     * @type {string}
     * @memberof UpdateServiceCommand
     */
    'serviceId'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateServiceCommand
     */
    'serviceName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateServiceCommand
     */
    'description'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateServiceCommand
     */
    'url'?: string | null;
    /**
     * 
     * @type {AuthMethod}
     * @memberof UpdateServiceCommand
     */
    'minimumAuthMethod'?: AuthMethod;
    /**
     * 
     * @type {ServiceType}
     * @memberof UpdateServiceCommand
     */
    'serviceType'?: ServiceType;
    /**
     * 
     * @type {string}
     * @memberof UpdateServiceCommand
     */
    'authorizeURL'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateServiceCommand
     */
    'logoutHookURL'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof UpdateServiceCommand
     */
    'feNotificationsEnabled'?: boolean;
}
// ...
export class ServicesManagerApi extends BaseAPI {
    /**
     * 
     * @param {CreateServiceCommand} [createServiceCommand] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ServicesManagerApi
     */
    public servicesManagerCreateService(createServiceCommand?: CreateServiceCommand, options?: AxiosRequestConfig) {
        return ServicesManagerApiFp(this.configuration).servicesManagerCreateService(createServiceCommand, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} serviceId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ServicesManagerApi
     */
    public servicesManagerDeleteService(serviceId: string, options?: AxiosRequestConfig) {
        return ServicesManagerApiFp(this.configuration).servicesManagerDeleteService(serviceId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} serviceId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ServicesManagerApi
     */
    public servicesManagerGetServiceDetail(serviceId: string, options?: AxiosRequestConfig) {
        return ServicesManagerApiFp(this.configuration).servicesManagerGetServiceDetail(serviceId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} [searchQuery] 
     * @param {number} [pageNumber] 
     * @param {number} [pageSize] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ServicesManagerApi
     */
    public servicesManagerGetServices(searchQuery?: string, pageNumber?: number, pageSize?: number, options?: AxiosRequestConfig) {
        return ServicesManagerApiFp(this.configuration).servicesManagerGetServices(searchQuery, pageNumber, pageSize, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {UpdateServiceCommand} [updateServiceCommand] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ServicesManagerApi
     */
    public servicesManagerUpdateService(updateServiceCommand?: UpdateServiceCommand, options?: AxiosRequestConfig) {
        return ServicesManagerApiFp(this.configuration).servicesManagerUpdateService(updateServiceCommand, options).then((request) => request(this.axios, this.basePath));
    }
}